+++ src/app.rs
@@ -17,7 +17,7 @@ pub enum Command {
     },
 
     /// Apply sequence of moves to a position
-    ApplySequence {
+    Apply {
         /// Hexchess state
         fen: String,
 
@@ -53,7 +53,7 @@ pub enum Command {
 pub fn handle(app: App) -> Result<String, String> {
     match app.command {
         Command::AllTargets { fen } => commands::all_targets::execute(fen),
-        Command::ApplySequence { fen, sequence } => commands::apply_sequence::execute(fen, sequence),
+        Command::Apply { fen, sequence } => commands::apply::execute(fen, sequence),
         Command::TestMove { fen, notation } => commands::test_move::execute(fen, notation),
         Command::GetTargets { fen, position } => commands::get_targets::execute(fen, position),
         Command::Parse { fen } => commands::parse::execute(fen),